+++ app/(routes)/(site)/blog/page.tsx
@@ -26,7 +26,7 @@ export async function generateMetadata({
 
 export default async function BlogPage({ searchParams }: Props) {
   const tag = (await searchParams).tag;
-  const posts = await fetchAllPosts({ tag });
+  const posts = await fetchAllPosts({ tag, publishedOnly: true });
 
   if (!posts.success) return <PlaceholderPage />;